boinc-client: Do not overwrite same file when using usrmerge
in case usrmerge is enabled, unit file is already in expected location

(cherry picked from commit b8e97f5b2a11c0d4be9f61821aead402780a4be7)

+++ meta-oe/recipes-extended/boinc/boinc-client_7.18.1.bb
@@ -76,7 +76,8 @@ do_compile:prepend () {
 }
 
 do_install:append() {
-	if [ -e ${D}${nonarch_libdir}/systemd/system/boinc-client.service ]; then
+	if ! ${@bb.utils.contains('DISTRO_FEATURES','usrmerge','true','false',d)} -a \
+            -e ${D}${nonarch_libdir}/systemd/system/boinc-client.service; then
 		install -d ${D}${systemd_system_unitdir}
 		mv \
 		${D}${nonarch_libdir}/systemd/system/boinc-client.service \
